+++ native/swift/Tests/integration-tests/MediaTests.swift
@@ -1,6 +1,7 @@
 import Foundation
-@testable import WordPressAPI
 import Testing
+import WordPressApiCache
+@testable import WordPressAPI
 
 @Suite(.serialized)
 struct MediaTests {
@@ -46,6 +47,26 @@ struct MediaTests {
             fulfilling: progress
         )
         #expect(response.data.mimeType == "image/jpeg")
+        #expect(progress.fractionCompleted == 1)
+
+        try await restoreTestServer()
+    }
+
+    @Test
+    func uploadProgressWithService() async throws {
+        let cache = try WordPressApiCache()
+        _ = try cache.performMigrations()
+        let service = try api.createService(cache: cache)
+
+        let progress = Progress.discreteProgress(totalUnitCount: 100)
+        #expect(progress.fractionCompleted == 0)
+
+        let file = try #require(Bundle.module.url(forResource: "test-data/test_media.jpg", withExtension: nil))
+        let response = try await service.uploadMedia(
+            params: .init(filePath: file.path),
+            fulfilling: progress
+        )
+        #expect(response.mimeType == "image/jpeg")
         #expect(progress.fractionCompleted == 1)
 
         try await restoreTestServer()